Checking in a working prototype
A fair bit more work is needed here, but before I update the underlying
mqtt library I wanted to check in the code. So far, the code is able to
login and get a list of supported devices. The payload is also able to
turn on and off Mss425e devices. The usage documentation will be updated
shortly.

Signed-off-by: Alex Boten <alrex.boten@gmail.com>